+++ sphinx/activity/main/activitymain/src/main/java/chat/sphinx/activitymain/navigation/drivers/DetailNavigationDriver.kt
@@ -6,7 +6,7 @@ import io.matthewnelson.android_feature_navigation.requests.PopBackStack
 import io.matthewnelson.concept_navigation.NavigationRequest
 import io.matthewnelson.feature_navigation.NavigationDriver
 
-internal class DetailNavigationDriver: NavigationDriver<NavController>(replayCacheSize = 2) {
+class DetailNavigationDriver: NavigationDriver<NavController>(replayCacheSize = 2) {
     override suspend fun whenTrueExecuteRequest(request: NavigationRequest<NavController>): Boolean {
         return true
     }