+++ sphinx/activity/main/activitymain/src/main/java/chat/sphinx/activitymain/navigation/navigators/primary/ProfileNavigatorImpl.kt
@@ -1,8 +1,10 @@
 package chat.sphinx.activitymain.navigation.navigators.primary
 
+import chat.sphinx.activitymain.R
 import chat.sphinx.activitymain.navigation.drivers.DetailNavigationDriver
 import chat.sphinx.activitymain.navigation.drivers.PrimaryNavigationDriver
 import chat.sphinx.example.manage_storage.navigation.ToManageStorageDetail
+import chat.sphinx.onboard_welcome.navigation.ToOnBoardWelcomeScreen
 import chat.sphinx.qr_code.navigation.ToQRCodeDetail
 import chat.sphinx.profile.navigation.ProfileNavigator
 import javax.inject.Inject
@@ -22,5 +24,11 @@ internal class ProfileNavigatorImpl @Inject constructor(
 
     override suspend fun toManageStorageDetail() {
         detailDriver.submitNavigationRequest(ToManageStorageDetail())
+    }
+
+    override suspend fun toOnBoardWelcomeScreen() {
+        navigationDriver.submitNavigationRequest(
+            ToOnBoardWelcomeScreen(popUpToId = R.id.main_primary_nav_graph)
+        )
     }
 }