+++ sphinx/application/sphinx/src/main/java/chat/sphinx/authentication/SphinxKeyRestore.kt
@@ -148,4 +148,10 @@ class SphinxKeyRestore(
             } ?: emit(KeyRestoreResponse.Error.InvalidUserPin)
         }
     }
+
+    override suspend fun clearAll() {
+        authenticationStorage.clearAuthenticationStorage()
+        authenticationManager.logOut()
+        encryptionKeyHandler.clearKeysToRestore()
+    }
 }