+++ sphinx/screens/onboard/key-restore/src/main/java/chat/sphinx/key_restore/KeyRestore.kt
@@ -20,4 +20,6 @@ abstract class KeyRestore {
         authorizationToken: AuthorizationToken,
         transportKey: RsaPublicKey? = null
     ): Flow<KeyRestoreResponse>
+
+    abstract suspend fun clearAll()
 }