+++ sphinx/screens/profile/profile/src/main/java/chat/sphinx/profile/ui/ProfileFragment.kt
@@ -279,6 +279,10 @@ internal class ProfileFragment: SideEffectFragment<
                     viewModel.resetPIN()
                 }
 
+                buttonProfileDeleteAccount.setOnClickListener {
+                    viewModel.deleteAccount()
+                }
+
                 includeProfileBasicContainerHolder.layoutConstraintProfileBasicContainerManageStorage.setOnClickListener {
                     lifecycleScope.launch(viewModel.mainImmediate) {
                         profileNavigator.toManageStorageDetail()