+++ sphinx/service/features/notifications/feature-service-notification-firebase/src/main/java/chat/sphinx/feature_service_notification_firebase/SphinxFirebaseMessagingService.kt
@@ -1,5 +1,12 @@
 package chat.sphinx.feature_service_notification_firebase
 
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.content.Intent
+import android.os.Build
+import androidx.core.app.NotificationCompat
+import chat.sphinx.activitymain.MainActivity
 import chat.sphinx.logger.SphinxLogger
 import chat.sphinx.logger.d
 import com.google.firebase.messaging.FirebaseMessagingService
@@ -8,6 +15,8 @@ import dagger.hilt.android.AndroidEntryPoint
 import io.matthewnelson.concept_coroutines.CoroutineDispatchers
 import kotlinx.coroutines.*
 import javax.inject.Inject
+import chat.sphinx.resources.R as R_common
+
 
 /*
  * https://firebase.google.com/docs/cloud-messaging/android/first-message
@@ -37,11 +46,63 @@ internal class SphinxFirebaseMessagingService: FirebaseMessagingService() {
 
     override fun onMessageReceived(p0: RemoteMessage) {
         super.onMessageReceived(p0)
-        LOG.d(TAG, "From: ${p0.from}")
-        LOG.d(TAG, "Notification Body: ${p0.notification?.body}")
-        for ((index, entry) in p0.data.entries.withIndex()) {
-            LOG.d(TAG, "Data: Index(value=$index), Key(value=${entry.key}), Value(value=${entry.value})")
+
+        if (p0.notification != null) {
+            ///Old notification, no need to create notification from code
+            return
         }
+
+        if (MainActivity.isActive) {
+            ///If app is in foreground, then do not show notification
+            return
+        }
+
+        // Extract data from the message
+        // Extract data from the message
+        val title: String = p0.data["message"] ?: ""
+        val messageBody: String = p0.data["body"] ?: ""
+        val chatId: Long? = (p0.data["chat_id"] ?: "").toLongOrNull()
+
+        // Create an intent to open MainActivity when the notification is clicked
+        val intent = Intent(
+            this,
+            MainActivity::class.java
+        )
+
+        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+        intent.putExtra("chat_id", chatId)
+
+        val pendingIntent = PendingIntent.getActivity(
+            this,
+            0,
+            intent,
+            PendingIntent.FLAG_ONE_SHOT or PendingIntent.FLAG_IMMUTABLE
+        )
+
+        // Build the notification
+        val notificationBuilder: NotificationCompat.Builder =
+            NotificationCompat.Builder(this, "channel_id")
+                .setSmallIcon(R_common.drawable.sphinx_white_logo)
+                .setContentTitle(title)
+                .setContentText(messageBody)
+                .setAutoCancel(true)
+                .setContentIntent(pendingIntent)
+
+        // Get the NotificationManager
+        val notificationManager = getSystemService(NOTIFICATION_SERVICE) as NotificationManager
+
+        // Check if the device is running Android Oreo or higher
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val channel = NotificationChannel(
+                "channel_id",
+                "Channel Name",
+                NotificationManager.IMPORTANCE_DEFAULT
+            )
+            notificationManager.createNotificationChannel(channel)
+        }
+
+        // Show the notification
+        notificationManager.notify(0, notificationBuilder.build())
     }
 
     override fun onNewToken(p0: String) {