+++ docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # ChangeLog
 
+# Version 1.0.0-beta13 (2023-12-13)
+- Adds ability to delete Sphinx account
+
+# Version 1.0.0-beta12 (2023-12-13)
+- Improves push notifications handling
+- Enables Create Tribe for virtual nodes
+
+# Version 1.0.0-beta11 (2023-11-17)
+- Fixes crashes
+- Fixes Out of Memory issues
+
 # Version 1.0.0-beta10 (2023-11-03)
 - Fixes issue on invoice tag